refactor: rename files and restructure namespace for command interfaces handling
