+++ tests/Fixtures/Commands/HelpStringDescriptionWithScopesCommand.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace SergiX44\Nutgram\Tests\Fixtures\Commands;
 
-use SergiX44\Nutgram\Handlers\Type\TelegramCommand;
-use SergiX44\Nutgram\Handlers\Type\WithScopes;
+use SergiX44\Nutgram\Handlers\Type\Command\Command;
+use SergiX44\Nutgram\Handlers\Type\Command\WithScopes;
 use SergiX44\Nutgram\Nutgram;
 use SergiX44\Nutgram\Telegram\Types\Command\BotCommandScopeAllPrivateChats;
 
-class HelpStringDescriptionWithScopesCommand implements TelegramCommand, WithScopes
+class HelpStringDescriptionWithScopesCommand implements Command, WithScopes
 {
     public function description(): string|array
     {
